+++ CHANGELOG.md
@@ -1,3 +1,53 @@
+## [1.4.1](https://github.com/akmatsu/Matanuska-Susitna-Web/compare/v1.4.0...v1.4.1) (2026-04-03)
+
+
+### Bug Fixes
+
+* fixed directory button text ([140090a](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/140090a28160d822690e9d80a78e2953a855dc40))
+
+# [1.4.0](https://github.com/akmatsu/Matanuska-Susitna-Web/compare/v1.3.0...v1.4.0) (2026-04-03)
+
+
+### Features
+
+* fixed assembly directory ([1535ef6](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/1535ef65322337cf023ebde167c8eda1c092e6fd))
+
+# [1.3.0](https://github.com/akmatsu/Matanuska-Susitna-Web/compare/v1.2.1...v1.3.0) (2026-04-03)
+
+
+### Features
+
+* updated language ([0968008](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/0968008c9935dd439e863e837c6d451faaa1c133))
+
+## [1.2.1](https://github.com/akmatsu/Matanuska-Susitna-Web/compare/v1.2.0...v1.2.1) (2026-04-03)
+
+
+### Bug Fixes
+
+* fixed board member term dates ([76a7d4c](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/76a7d4c8096094507a781de23a40d3c1a45cfd0c))
+
+# [1.2.0](https://github.com/akmatsu/Matanuska-Susitna-Web/compare/v1.1.0...v1.2.0) (2026-04-03)
+
+
+### Bug Fixes
+
+* fixed small UI issues in board directory display ([2893d3b](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/2893d3bfc6ce2b3e832c861dd6a8795e9e6ccf2a))
+
+
+### Features
+
+* added a quick link to board directory in search page ([818b5a9](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/818b5a95300b52121347d0304d97c387d84d1dc9))
+* board list table directory link now points to board directory page if available ([df5ce79](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/df5ce79ecee9f2275055820560fdbaed531306a9))
+* created board directory page ([994cfe7](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/994cfe77f97e0b1dd72962bf23240056da0c3134))
+* only display old directory if directory excel does not exist ([f1e5045](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/f1e504549b8bc4d52df870eecf83c8dc74999d7d))
+* only enabling page view tracker in production ([cd681f9](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/cd681f988598aebc075d82254e57e54c6671a32f))
+* only loading analytics script in production ([9090846](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/9090846db6c5ee9063f0cfce50082306068c6fec))
+* styling board directory; ([0929ccb](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/0929ccb8cd1f5e5f6c444cb17d262045616b3416))
+* updated board directory display ([8785c7b](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/8785c7b5051921a6382be3d8c259970cd0d4eeba))
+* updated deps and removed maps widget ([b571fe6](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/b571fe6a38f8c325c65db44c8cb79ced9c7bba95))
+* updated directory cards ([ad03340](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/ad03340f041f0d71c6d0766fd9d28b597d0011c9))
+* updated directory cards ([30b4a08](https://github.com/akmatsu/Matanuska-Susitna-Web/commit/30b4a0870b5cde751395bf316d61f04d7de03070))
+
 # [1.1.0](https://github.com/akmatsu/Matanuska-Susitna-Web/compare/v1.0.0...v1.1.0) (2026-03-20)